requirements.txt
click==8.1.3
faiss-cpu==1.7.0
Flask==2.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
langdetect==1.0.9
MarkupSafe==2.1.1
nltk==3.7
numpy==1.23.0
pandas==1.4.3
python-dateutil==2.8.2
pytz==2022.1
regex==2022.6.2
six==1.16.0
torch==1.11.0
tqdm==4.64.0
typing_extensions==4.2.0
Werkzeug==2.1.2
zipp==3.8.0
